app/interfaces/schemas/token.py
from pydantic import BaseModel

class RefreshTokenRequest(BaseModel):
    refresh_token: str

class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    refresh_token: str | None = None
    user_info: dict | None = None
