requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pyodbc
python-dotenv
requests
python-jose
passlib[bcrypt]
pydantic-settings>=2.0.0
python-multipart
msal
openpyxl
email-validator
